fix: move effects to within the components
