+++ packages/governance-app/src/Council/CouncilCandidates.tsx
@@ -1,0 +1,50 @@
+// Copyright 2018-2019 @paritytech/substrate-light-ui authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import { AccountId, Vector } from '@polkadot/types';
+import { AppContext } from '@substrate/ui-common';
+import { FadedText, Header, Stacked, StyledNavLink } from '@substrate/ui-components';
+import React, { useContext, useEffect, useState } from 'react';
+import { Observable } from 'rxjs';
+
+export function CouncilCandidates () {
+  const { api } = useContext(AppContext);
+  const [councilCandidates, setCouncilCandidates] = useState();
+
+  useEffect(() => {
+    const subscription =
+      (api.query.council.candidates() as unknown as Observable<Vector<AccountId>>)
+      .subscribe((councilCandidates) => {
+        setCouncilCandidates(councilCandidates);
+      });
+
+    return () => subscription.unsubscribe();
+  }, []);
+
+  const renderCouncilCandidates = () => {
+    if (councilCandidates && councilCandidates.length) {
+      return (
+        <React.Fragment>
+          <Header>Council Candidates ({councilCandidates && councilCandidates.length})</Header>
+          {
+            councilCandidates.map((accountId: AccountId) => {
+              return (
+                <li>{accountId} is a candidate</li>
+              );
+            })
+          }
+        </React.Fragment>
+      )
+    } else {
+      return (
+        <Stacked>
+          <FadedText>No Candidates Found</FadedText>
+          <StyledNavLink to={'/#'}>Click here to learn more about the Council</StyledNavLink>
+        </Stacked>
+      );
+    }
+  };
+
+  return renderCouncilCandidates();
+}
